+++ node/src/main/scala/org/bitcoins/node/NeutrinoNode.scala
@@ -13,10 +13,9 @@ import org.bitcoins.core.api.chain.db.{
 import org.bitcoins.core.protocol.BlockStamp
 import org.bitcoins.node.config.NodeAppConfig
 import org.bitcoins.node.models.Peer
-import org.bitcoins.node.networking.peer.{DataMessageHandler, PeerMessageSender}
+import org.bitcoins.node.networking.peer.DataMessageHandler
 
 import scala.concurrent.Future
-import scala.util.Random
 
 case class NeutrinoNode(
     nodePeer: Vector[Peer],
@@ -40,17 +39,6 @@ case class NeutrinoNode(
   override def updateDataMessageHandler(
       dataMessageHandler: DataMessageHandler): NeutrinoNode = {
     copy(dataMessageHandler = dataMessageHandler)
-  }
-
-  lazy val randomPeerMsgSenderWithCompactFilters: PeerMessageSender = {
-    val filteredPeers =
-      peerServices.filter(_._2.nodeCompactFilters).keys.toVector
-    if (filteredPeers.isEmpty)
-      throw new RuntimeException("No peers supporting compact filters!")
-    val peer = filteredPeers(Random.nextInt(filteredPeers.length))
-    peerMsgSenders
-      .find(_.client.peer == peer)
-      .getOrElse(throw new RuntimeException("This should not happen."))
   }
 
   override def start(): Future[NeutrinoNode] = {
@@ -77,7 +65,6 @@ case class NeutrinoNode(
     * @return
     */
   override def sync(): Future[Unit] = {
-    val peerMsgSender = randomPeerMsgSenderWithCompactFilters
     val blockchainsF =
       BlockHeaderDAO()(executionContext, chainConfig).getBlockchains()
     for {
@@ -88,7 +75,7 @@ case class NeutrinoNode(
       blockchains <- blockchainsF
       // Get all of our cached headers in case of a reorg
       cachedHeaders = blockchains.flatMap(_.headers).map(_.hashBE.flip)
-      _ <- peerMsgSender.sendGetHeadersMessage(cachedHeaders)
+      _ <- randomPeerMsgSender.sendGetHeadersMessage(cachedHeaders)
       _ <- syncFilters(bestFilterHeaderOpt = bestFilterHeaderOpt,
                        bestFilterOpt = bestFilterOpt,
                        bestBlockHeader = header,
@@ -141,10 +128,11 @@ case class NeutrinoNode(
       chainApi: ChainApi,
       bestFilterOpt: Option[CompactFilterDb]): Future[Unit] = {
     val sendCompactFilterHeaderMsgF = {
-      peerMsgSenders(0).sendNextGetCompactFilterHeadersCommand(
-        chainApi = chainApi,
-        filterHeaderBatchSize = chainConfig.filterHeaderBatchSize,
-        prevStopHash = bestFilterHeader.blockHashBE)
+      randomPeerMsgSenderWithCompactFilters
+        .sendNextGetCompactFilterHeadersCommand(
+          chainApi = chainApi,
+          filterHeaderBatchSize = chainConfig.filterHeaderBatchSize,
+          prevStopHash = bestFilterHeader.blockHashBE)
     }
     sendCompactFilterHeaderMsgF.flatMap { isSyncFilterHeaders =>
       // If we have started syncing filters
@@ -156,7 +144,7 @@ case class NeutrinoNode(
         //means we are not syncing filter headers, and our filters are NOT
         //in sync with our compact filter headers
         logger.info(s"Starting sync filters in NeutrinoNode.sync()")
-        peerMsgSenders(0)
+        randomPeerMsgSenderWithCompactFilters
           .sendNextGetCompactFilterCommand(
             chainApi = chainApi,
             filterBatchSize = chainConfig.filterBatchSize,
